FE: Added components: TinyMCE wrapper, VisibilitySelect, AlertComponent. addPost data logic moved to container.
